+++ src/Controllers/balanceControllers.ts
@@ -1,6 +1,6 @@
-import { Expense as ExpenseModel } from '../../models/Expense';
-import { Balance as BalanceModel } from '../../models/Balance';
-import {Income as IncomeModel} from '../../models/Income';
+import { Expense as ExpenseModel } from '../models/Expense';
+import { Balance as BalanceModel } from '../models/Balance';
+import {Income as IncomeModel} from '../models/Income';
 
 export const getAllMovements = async (id: number) => {
