+++ src/Handlers/Classificators/Filters/incomeFilterHandler.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
-import { incomeDateFilter, incomeAccountFilter } from '../../Controllers/incomeController/filterController';
-import { IDate } from '../expenseClassification/filterHandler';
+import { incomeDateFilter, incomeAccountFilter } from '../../../Controllers/Classificators/Filters/incomeFilterController';
+import { IDate } from './expenseFilterHandler';
 
 export const getIncomeDateFilter = async(req: Request, res: Response) =>{
     const date: IDate= req.body;
@@ -14,8 +14,8 @@ export const getIncomeDateFilter = async(req: Request, res: Response) =>{
     } catch (error) {
         console.error('Error ocurred while filter incomes...', error);
     res
-      .status(400)
-      .json({ message: 'Failed to filter incomes by accounts. Try again later...' });
+        .status(400)
+        .json({ message: 'Failed to filter incomes by accounts. Try again later...' });
 
     }
 }
@@ -23,7 +23,7 @@ export const getIncomeDateFilter = async(req: Request, res: Response) =>{
 export interface IAccount{
     account:string
 };
- 
+
 export const getIncomeAccountFilter = async(req: Request, res: Response) =>{
     const account: IAccount = req.body;
     const { id } = req.params;
@@ -37,8 +37,8 @@ export const getIncomeAccountFilter = async(req: Request, res: Response) =>{
     } catch (error) {
         console.error('Error ocurred while filter incomes...', error);
     res
-      .status(400)
-      .json({ message: 'Failed to filter incomes by accounts. Try again later...' });
+        .status(400)
+        .json({ message: 'Failed to filter incomes by accounts. Try again later...' });
 
     }
 }
