+++ src/models/Expense.ts
@@ -7,7 +7,7 @@ import {
   ForeignKey,
   BelongsTo
 } from 'sequelize-typescript';
-import { Accounts } from './Accounts';
+import { Account } from './Account';
 
 @Table
 export class Expense extends Model<Expense> {
@@ -46,10 +46,10 @@ export class Expense extends Model<Expense> {
   @Column
   updatedAt!: Date;
 
-  @ForeignKey(() => Accounts)
+  @ForeignKey(() => Account)
   @Column
   accountId!: number;
 
-  @BelongsTo(() => Accounts)
-  Accounts!: Accounts;
+  @BelongsTo(() => Account)
+  Account!: Account;
 }
