+++ src/routes/user.routes.ts
@@ -1,10 +1,13 @@
 import { Router } from 'express';
 import { getUser, getUsers, postUser, putUser } from '../Handlers/usersHandler';
+
 const userRouter: Router = Router();
 
-userRouter.get('/', getUsers);
-userRouter.get('/:id', getUser);
-userRouter.post('/', postUser);
-userRouter.put('/:id', putUser);
+    userRouter.get('/', getUsers);
+    userRouter.get('/:id', getUser);
+
+    userRouter.post('/', postUser);
+
+    userRouter.put('/:id', putUser);
 
 export default userRouter;
